pyproject.toml
[project]
name = "sonos-lastfm"
version = "0.1.0"
description = "Scrobble songs played on Sonos speakers to Last.fm"
requires-python = ">=3.14"
dependencies = [
    "soco>=0.30",
    "pylast>=5.3",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "python-json-logger>=4.0.0",
]

[project.scripts]
sonos-lastfm = "sonos_lastfm.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sonos_lastfm"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.14",
]

[tool.coverage.run]
omit = ["src/sonos_lastfm/__main__.py"]
